chore: move all redteam constants to one spot
