+++ src/app/src/pages/redteam/dashboard/components/CategoryBreakdown.tsx
@@ -1,10 +1,4 @@
 import React from 'react';
-import {
-  categoryAliases,
-  riskCategories,
-  displayNameOverrides,
-  subCategoryDescriptions,
-} from '@app/pages/redteam/report/components/constants';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
 import Table from '@mui/material/Table';
@@ -14,6 +8,12 @@ import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
+import {
+  categoryAliases,
+  riskCategories,
+  displayNameOverrides,
+  subCategoryDescriptions,
+} from '@promptfoo/redteam/constants';
 import type { StandaloneEval } from '@promptfoo/util';
 import { processCategoryData } from './utils';
