+++ src/app/src/pages/redteam/dashboard/components/TopFailingCategories.tsx
@@ -1,15 +1,12 @@
 import React from 'react';
 import type { StandaloneEval } from '@/../../../util';
-import {
-  categoryAliases,
-  displayNameOverrides,
-} from '@app/pages/redteam/report/components/constants';
 import Box from '@mui/material/Box';
 import LinearProgress from '@mui/material/LinearProgress';
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
 import Typography from '@mui/material/Typography';
+import { categoryAliases, displayNameOverrides } from '@promptfoo/redteam/constants';
 import { processCategoryData } from './utils';
 
 interface TopFailingCategoriesProps {
