+++ src/app/src/pages/redteam/report/components/RiskCategories.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Stack from '@mui/material/Stack';
+import type { TopLevelCategory } from '@promptfoo/redteam/constants';
+import { categoryDescriptions, riskCategories } from '@promptfoo/redteam/constants';
 import RiskCard from './RiskCard';
-import type { TopLevelCategory } from './constants';
-import { categoryDescriptions, riskCategories } from './constants';
 
 const RiskCategories: React.FC<{
   categoryStats: Record<string, { pass: number; total: number }>;
