+++ src/app/src/pages/redteam/report/components/RiskCategoryDrawer.tsx
@@ -7,9 +7,9 @@ import IconButton from '@mui/material/IconButton';
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import Typography from '@mui/material/Typography';
+import { categoryAliases, displayNameOverrides } from '@promptfoo/redteam/constants';
 import type { GradingResult } from '@promptfoo/types';
 import SuggestionsDialog from './SuggestionsDialog';
-import { categoryAliases, displayNameOverrides } from './constants';
 import './RiskCategoryDrawer.css';
 
 interface RiskCategoryDrawerProps {
